separate irc & tg functionality into own sub-directories
