+++ src/tg/nickcolor.js
